backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
motor         
pymongo  